pyproject.toml
[project]
name = "ads-worker-framework"
version = "0.1.0"
description = "AdsPower multi-environment automation framework with Playwright"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = []
